Rename ow_synchronize.h to ow_abilities.h (source name was already changed)
